doc: Move OverlayLayer from containers/ to components/
