+++ apps/insight-viewer-demo/src/containers/Overlay/index.tsx
@@ -1,7 +1,7 @@
 import { Box } from '@chakra-ui/react'
 import Viewer, { useViewport } from '@lunit/insight-viewer'
 import CodeBlock from '../../components/CodeBlock'
-import OverlayLayer from './OverlayLayer'
+import OverlayLayer from '../../components/OverlayLayer'
 
 const IMAGE_ID =
   'wadouri:https://static.lunit.io/fixtures/dcm-files/series/CT000011.dcm'
